skal flytte logikk for å sende endringsmelding fra send endringsmelding til dokumentasjon
