+++ app/sider/SendEndring.tsx
@@ -1,15 +1,9 @@
-import { Alert, Button, Textarea } from '@navikt/ds-react';
-import {
-  Form,
-  useActionData,
-  useNavigate,
-  useNavigation,
-  useSubmit,
-} from '@remix-run/react';
-import React, { useEffect, useState } from 'react';
+import { Button, Textarea } from '@navikt/ds-react';
+import { useNavigate, useNavigation } from '@remix-run/react';
+import React, { useState } from 'react';
 
 import {
-  useEndringsmeldingMottattDato,
+  useEndringsmelding,
   useSpråk,
   useTekster,
   useYtelse,
@@ -19,13 +13,9 @@ import Spinner from '~/komponenter/spinner/Spinner';
 import StegIndikator from '~/komponenter/stegindikator/StegIndikator';
 import TekstBlokk from '~/komponenter/tekstblokk/TekstBlokk';
 import VeilederPanel from '~/komponenter/veilederpanel/VeilederPanel';
-import { action as actionBA } from '~/routes/ba.endringsmelding';
-import { action as actionKS } from '~/routes/ks.endringsmelding';
 import { ESanityMappe, ESteg } from '~/typer/felles';
 import { EFritekstFeil, fritekstFeilTilApiKeys } from '~/typer/fritekstfeil';
-import { EStatusKode, IPostResponse } from '~/typer/response';
 import { ETypografiTyper } from '~/typer/typografi';
-import { EYtelse } from '~/typer/ytelse';
 import {
   i18nInnhold,
   MAKS_INPUT_LENGDE,
@@ -37,33 +27,18 @@ import css from './send-endringsmelding.module.css';
 
 export default function SendEndringSide() {
   const ytelse = useYtelse();
-  const action = ytelse === EYtelse.BARNETRYGD ? actionBA : actionKS;
-  const actionData: IPostResponse | undefined = useActionData<typeof action>();
-  const submit = useSubmit();
-
   const tekster = useTekster(ESanityMappe.SEND_ENDRINGER);
   const teksterFelles = useTekster(ESanityMappe.FELLES);
   const navigate = useNavigate();
   const navigation = useNavigation();
   const [språk] = useSpråk();
-  const [, settEndringsmeldingMottattDato] = useEndringsmeldingMottattDato();
-
-  const [feilKode, settFeilKode] = useState<EFritekstFeil | null>(null);
+  const [, settEndringsmelding] = useEndringsmelding();
   const [valideringsfeil, settValideringsfeil] = useState<EFritekstFeil | null>(
     EFritekstFeil.MANGLER_TEKST,
   );
   const [erKnappTrykketPå, settKnappTrykketPå] = useState<boolean>(false);
 
-  useEffect(() => {
-    if (!actionData) return;
-
-    if (actionData.status === EStatusKode.OK && actionData.data) {
-      settEndringsmeldingMottattDato(actionData.data.mottattDato);
-      navigate(hentPathForSteg(ytelse, ESteg.KVITTERING));
-    } else {
-      settFeilKode(actionData.feilKode || null);
-    }
-  }, [actionData, navigate, settEndringsmeldingMottattDato, ytelse]);
+  const [endringsmeldingInput, settEndringsmeldingInput] = useState<string>('');
 
   function genererFeilmelding() {
     return (
@@ -77,13 +52,10 @@ export default function SendEndringSide() {
     );
   }
 
-  function håndterSendEndringsmelding(
-    event: React.MouseEvent<HTMLButtonElement, MouseEvent>,
-  ) {
-    event.preventDefault();
+  function håndterTrykkNeste() {
     if (valideringsfeil === null) {
-      settFeilKode(null);
-      submit(event.currentTarget, { replace: true });
+      settEndringsmelding({ tekst: endringsmeldingInput, dokumenter: [] });
+      navigate(hentPathForSteg(ytelse, ESteg.DOKUMENTASJON));
     }
     settKnappTrykketPå(true);
   }
@@ -101,64 +73,44 @@ export default function SendEndringSide() {
             dataTestid="overskriftSteg1"
           />
           <VeilederPanel innhold={tekster.veilederInnhold} />
-          <Form method="post" className={`${css.fullBredde}`}>
-            <Textarea
-              data-testid="fritekstfelt"
-              name="endringsmelding"
-              label={
-                <TekstBlokk
-                  tekstblokk={tekster.fritekstfeltTittel}
-                  dataTestid="fritekstfeltTittel"
-                />
-              }
-              description={
-                <TekstBlokk tekstblokk={tekster.fritekstfeltBeskrivelse} />
-              }
-              autoComplete="on"
-              maxLength={MAKS_INPUT_LENGDE}
-              i18n={i18nInnhold(språk)}
-              error={genererFeilmelding()}
-              onInput={event => {
-                settValideringsfeil(validerTekst(event.currentTarget.value));
-              }}
-            />
-            {feilKode && (
-              <Alert variant="error" className={`${css.toppMargin}`}>
-                <TekstBlokk
-                  tekstblokk={tekster[fritekstFeilTilApiKeys[feilKode]]}
-                />
-              </Alert>
-            )}
-            <div className={`${css.navigeringsKnappKonteiner}`}>
-              <Button
-                type="button"
-                variant={'secondary'}
-                onClick={() => navigate(hentPathForSteg(ytelse, ESteg.FORSIDE))}
-              >
-                <TekstBlokk tekstblokk={teksterFelles.knappTilbake} />
-              </Button>
-              <Button
-                type="button"
-                variant={valideringsfeil === null ? 'primary' : 'secondary'}
-                data-testid="knappVidereSteg2"
-                onClick={() =>
-                  navigate(hentPathForSteg(ytelse, ESteg.DOKUMENTASJON))
-                }
-              >
-                <TekstBlokk tekstblokk={teksterFelles.knappNeste} />
-              </Button>
-              <Button
-                type="submit"
-                variant={valideringsfeil === null ? 'primary' : 'secondary'}
-                data-testid="knappVidereSteg1"
-                onClick={event => {
-                  håndterSendEndringsmelding(event);
-                }}
-              >
-                <TekstBlokk tekstblokk={teksterFelles.knappSendEndringer} />
-              </Button>
-            </div>
-          </Form>
+          <Textarea
+            data-testid="fritekstfelt"
+            name="endringsmelding"
+            label={
+              <TekstBlokk
+                tekstblokk={tekster.fritekstfeltTittel}
+                dataTestid="fritekstfeltTittel"
+              />
+            }
+            description={
+              <TekstBlokk tekstblokk={tekster.fritekstfeltBeskrivelse} />
+            }
+            autoComplete="on"
+            maxLength={MAKS_INPUT_LENGDE}
+            i18n={i18nInnhold(språk)}
+            error={genererFeilmelding()}
+            onInput={event => {
+              settValideringsfeil(validerTekst(event.currentTarget.value));
+              settEndringsmeldingInput(event.currentTarget.value);
+            }}
+          />
+          <div className={`${css.navigeringsKnappKonteiner}`}>
+            <Button
+              type="button"
+              variant={'secondary'}
+              onClick={() => navigate(hentPathForSteg(ytelse, ESteg.FORSIDE))}
+            >
+              <TekstBlokk tekstblokk={teksterFelles.knappTilbake} />
+            </Button>
+            <Button
+              type="button"
+              variant={valideringsfeil === null ? 'primary' : 'secondary'}
+              data-testid="knappVidereSteg2"
+              onClick={håndterTrykkNeste}
+            >
+              <TekstBlokk tekstblokk={teksterFelles.knappNeste} />
+            </Button>
+          </div>
         </>
       )}
     </HovedInnhold>
