+++ app/sider/dokumentasjon.module.css
@@ -6,3 +6,7 @@
   justify-content: center;
   gap: 1rem;
 }
+
+.toppMargin {
+  margin-top: 5rem;
+}
