+++ app/sider/send-endringsmelding.module.css
@@ -6,11 +6,3 @@
   justify-content: center;
   gap: 1rem;
 }
-
-.fullBredde {
-  width: 100%;
-}
-
-.toppMargin {
-  margin-top: 5rem;
-}
